backend/requirements.txt
fastapi
uvicorn
tensorflow
pillow
sqlalchemy
python-multipart
python-jose[cryptography]
passlib[bcrypt]
